requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.13
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1